+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [3.12.0] - 2026-06-08
 
+* Add "capacity" availability timeout
+
 * Add "max headers" check, default is 96 headers
 
 ## [3.11.1] - 2026-05-05